requirements.txt
flask==2.0.1
werkzeug==2.0.3
python-dotenv==0.19.0
google-generativeai==0.3.0
firebase-admin==6.2.0
